requirements-api.txt
# Lightweight dependencies for the API server (no PyTorch / ML models)
fastapi
uvicorn
pydantic==2.5.0
python-multipart
python-dotenv
redis[hiredis]
requests
supabase
omise
aiohttp
Pillow
httpx==0.27.2
sentry-sdk[fastapi]
